+++ datarobot_batch_scoring/main.py
@@ -58,7 +58,6 @@ def parse_args(argv, standalone=False):
         'stdout': False,
         'auto_sample': False,
         'api_version': PRED_API_V10,
-        'skip_verify_ssl': False
     }
     parser = argparse.ArgumentParser(
         description=DESCRIPTION, epilog=EPILOG,
@@ -163,12 +162,23 @@ def parse_args(argv, standalone=False):
                          default=False,
                          help='Compress batch. This can improve throughout '
                               'when bandwidth is limited.')
-    conn_gr.add_argument('--skip_verify_ssl', action='store_true',
-                         default=defaults['skip_verify_ssl'],
-                         help='Disable SSL certificates verification for '
-                              'HTTPS endpoints. By default SSL verification '
-                              'is enabled and batch_scoring throw SSLError '
-                              'if it\'s unable to verify the certificate')
+    conn_gr.add_argument('--verify_ssl',
+                         const=True,
+                         nargs='?',
+                         dest='verify_ssl',
+                         default=True,
+                         help='Enable SSL certificates verification for '
+                              'HTTPS endpoints. You can either omit the '
+                              'value to enable SSL verification or '
+                              'pass a path to CA bundle file as a value '
+                              'to use it for verification using private '
+                              'certificates. '
+                              'By default SSL verification is enabled.')
+    conn_gr.add_argument('--no_verify_ssl',
+                         action='store_false',
+                         dest='verify_ssl',
+                         help='Skip SSL certificates verification for HTTPS '
+                              'endpoints.')
     csv_gr = parser.add_argument_group('CVS parameters')
     csv_gr.add_argument('--keep_cols', type=str,
                         nargs='?',
@@ -309,8 +319,6 @@ def parse_generic_options(parsed_args):
     if not os.path.exists(dataset):
         ui.fatal('file {} does not exist.'.format(dataset))
 
-    verify_ssl = not parsed_args['skip_verify_ssl']
-
     ui.debug('batch_scoring v{}'.format(__version__))
 
     return {
@@ -333,7 +341,7 @@ def parse_generic_options(parsed_args):
         'skip_dialect': skip_dialect,
         'skip_row_id': skip_row_id,
         'timeout': timeout,
-        'verify_ssl': verify_ssl
+        'verify_ssl': parsed_args['verify_ssl'],
     }